+++ src/io/poll/mod.rs
@@ -1,0 +1,11 @@
+use cfg_if::cfg_if;
+
+cfg_if! {
+    if #[cfg(target_os = "macos")] {
+        mod macos;
+        pub(crate) use macos::*;
+    } else if #[cfg(unix)] {
+        mod unix;
+        pub(crate) use unix::*;
+    }
+}
