--- src/os/macos/mod.rs
@@ -1,2 +1,0 @@
-mod poll;
-pub(crate) use poll::*;
