--- src/os/mod.rs
@@ -1,10 +1,0 @@
-#[cfg(target_os = "macos")]
-mod macos;
-#[cfg(target_os = "macos")]
-pub(crate) use macos::*;
-#[cfg(all(unix, not(target_os = "macos")))]
-mod unix;
-#[cfg(all(unix, not(target_os = "macos")))]
-pub(crate) use unix::*;
-#[cfg(unix)]
-mod unix_common;
