+++ src/quirks.rs
@@ -2,7 +2,7 @@ use std::env;
 use std::io::{self, Write};
 use std::sync::OnceLock;
 
-pub(super) fn terminal_quirks_from_env() -> TerminalQuirks {
+pub(crate) fn terminal_quirks_from_env() -> TerminalQuirks {
     // This OnceLock is not here for efficiency, it's here so that
     // we have consistent results in case a consumer uses `set_var`.
     static TERMINAL_QUIRK: OnceLock<TerminalQuirks> = OnceLock::new();
@@ -54,18 +54,18 @@ fn terminal_quirk_from_env_eager() -> TerminalQuirks {
 }
 
 #[derive(Debug, Copy, Clone)]
-pub(super) enum TerminalQuirks {
+pub(crate) enum TerminalQuirks {
     None,
     Unsupported,
     Urxvt,
 }
 
 impl TerminalQuirks {
-    pub(super) fn is_known_unsupported(self) -> bool {
+    pub(crate) fn is_known_unsupported(self) -> bool {
         matches!(self, TerminalQuirks::Unsupported)
     }
 
-    pub(super) fn string_terminator(self) -> &'static [u8] {
+    pub(crate) fn string_terminator(self) -> &'static [u8] {
         const ST: &[u8] = b"\x1b\\";
         const BEL: u8 = 0x07;
 
@@ -79,11 +79,11 @@ impl TerminalQuirks {
         }
     }
 
-    pub(super) fn write_all(self, w: &mut dyn Write, bytes: &[u8]) -> io::Result<()> {
+    pub(crate) fn write_all(self, w: &mut dyn Write, bytes: &[u8]) -> io::Result<()> {
         w.write_all(bytes)
     }
 
-    pub(super) fn write_string_terminator(self, writer: &mut dyn Write) -> io::Result<()> {
+    pub(crate) fn write_string_terminator(self, writer: &mut dyn Write) -> io::Result<()> {
         self.write_all(writer, self.string_terminator())
     }
 }
